Correct statistical value calculation - now results are Double.

+++ TSNid/src/main/java/ua/edu/znu/tsnid/benchmark/LatencyStats.kt
@@ -6,29 +6,29 @@ import kotlin.math.sqrt
  * Statistics holder for latency measurements.
  */
 data class LatencyStats(
-    val mean: Long,
-    val stdDev: Long,
-    val ci95Lower: Long,
-    val ci95Upper: Long
+    val mean: Double,
+    val stdDev: Double,
+    val ci95Lower: Double,
+    val ci95Upper: Double
 )
 
 /**
  * Calculates mean, standard deviation, and 95% confidence interval for a list of values.
  */
 fun calculateStats(values: List<Long>): LatencyStats {
-    val mean = values.average().toLong()
+    val mean = values.average()
     val n = values.size
     
     // Sample variance: Σ(x - mean)² / (n - 1)
     val variance = if (n > 1) {
-        values.sumOf { (it - mean) * (it - mean) } / (n - 1).toDouble()
+        values.sumOf { (it - mean) * (it - mean) } / (n - 1)
     } else {
         0.0
     }
-    val stdDev = sqrt(variance).toLong()
+    val stdDev = sqrt(variance)
     
     // 95% CI using t-distribution approximation (t ≈ 1.984 for n=100)
-    val marginOfError = (1.984 * stdDev / sqrt(n.toDouble())).toLong()
+    val marginOfError = 1.984 * stdDev / sqrt(n.toDouble())
     
     return LatencyStats(
         mean = mean,
